split Toolbox into multiple files
